// src/api/account/models/getSelfResponseContent.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * AccountService
 * OpenAPI spec version: 1.0.0
 */
import type { AccountRole } from "./accountRole";

export interface GetSelfResponseContent {
	id: number;
	name: string;
	role: AccountRole;
	created_at: string;
	updated_at: string;
}